In Nearby, only show checkmark if object is logged.
This conserves space in the screen, because most objects will not be
logged, the user is only interested to see which have been logged.
This also solves the case where some have empty checkboxes and some
Lines have no checkboxes because their width exceeded the space.
Now we show checked checkboxes even if the width is exceeded,
effectively removing size which is the biggest space hog anyway.
Rounding the sizes to arcsecond is a TODO, but that needs an overhaul so
that all sizes are recorded in the database in a uniform way.